requirements.txt
appdirs==1.4.3
boto3==1.12.37
botocore==1.15.37
docutils==0.15.2
fs==2.4.11
fs-s3fs==1.1.1
jmespath==0.9.5
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
